+++ pypeit/core/datacube.py
@@ -17,16 +17,12 @@ from scipy.interpolate import interp1d
 import numpy as np
 
 from pypeit import msgs
-from pypeit import spec2dobj, alignframe, masterframe
-from pypeit.core.flux_calib import load_extinction_data, extinction_correction
+from pypeit import alignframe, datamodel, flatfield, io, masterframe, specobj, spec2dobj, utils
+from pypeit.core.flux_calib import load_extinction_data, extinction_correction, fit_zeropoint, get_standard_spectrum, ZP_UNIT_CONST, PYPEIT_FLUX_SCALE
 from pypeit.core.flexure import calculate_image_offset
-from pypeit.core import parse
-#from pypeit.core.procimg import grow_masked
+from pypeit.core import coadd, extract, findobj_skymask, parse, skysub
 from pypeit.core.procimg import grow_mask
-from pypeit.core import coadd
 from pypeit.spectrographs.util import load_spectrograph
-from pypeit import datamodel
-from pypeit import io
 
 from IPython import embed
 
@@ -42,6 +38,10 @@ class DataCube(datamodel.DataContainer):
             The science datacube (nwave, nspaxel_y, nspaxel_x)
         variance (`numpy.ndarray`_):
             The variance datacube (nwave, nspaxel_y, nspaxel_x)
+        blaze_wave (`numpy.ndarray`_):
+            Wavelength array of the spectral blaze function
+        blaze_spec (`numpy.ndarray`_):
+            The spectral blaze function
         PYP_SPEC (str):
             Name of the PypeIt Spectrograph
         fluxed (bool):
@@ -58,10 +58,12 @@ class DataCube(datamodel.DataContainer):
             Build from PYP_SPEC
 
     """
-    version = '1.0.1'
+    version = '1.0.2'
 
     datamodel = {'flux': dict(otype=np.ndarray, atype=np.floating, descr='Flux array in units of counts/s/Ang or 10^-17 erg/s/cm^2/Ang'),
                  'variance': dict(otype=np.ndarray, atype=np.floating, descr='Variance array (matches units of flux)'),
+                 'blaze_wave': dict(otype=np.ndarray, atype=np.floating, descr='Wavelength array of the spectral blaze function'),
+                 'blaze_spec': dict(otype=np.ndarray, atype=np.floating, descr='The spectral blaze function'),
                  'PYP_SPEC': dict(otype=str, descr='PypeIt: Spectrograph name'),
                  'fluxed': dict(otype=bool, descr='Boolean indicating if the datacube is fluxed.')}
 
@@ -85,7 +87,7 @@ class DataCube(datamodel.DataContainer):
         slf.spect_meta = slf.spectrograph.parse_spec_header(slf.head0)
         return slf
 
-    def __init__(self, flux, variance, PYP_SPEC, fluxed=None):
+    def __init__(self, flux, variance, PYP_SPEC, blaze_wave, blaze_spec, fluxed=None):
 
         args, _, _, values = inspect.getargvalues(inspect.currentframe())
         _d = dict([(k,values[k]) for k in args[1:]])
@@ -265,6 +267,228 @@ def dar_correction(wave_arr, coord, obstime, location, pressure, temperature, re
     dec_diff = spl_dec(wave_arr) - spl_dec(wave_ref)
 
     return ra_diff, dec_diff
+
+
+def gaussian2D_cube(tup, amplitude, xo, yo, dxdz, dydz, sigma_x, sigma_y, theta, offset):
+    (x, y, z) = tup
+    xo = float(xo) + z*dxdz
+    yo = float(yo) + z*dydz
+    a = (np.cos(theta)**2)/(2*sigma_x**2) + (np.sin(theta)**2)/(2*sigma_y**2)
+    b = -(np.sin(2*theta))/(4*sigma_x**2) + (np.sin(2*theta))/(4*sigma_y**2)
+    c = (np.sin(theta)**2)/(2*sigma_x**2) + (np.cos(theta)**2)/(2*sigma_y**2)
+    norm = 1/(2*np.pi*np.sqrt(a*c-b*b))
+    g = offset + norm*amplitude*np.exp( - (a*((x-xo)**2) + 2*b*(x-xo)*(y-yo)
+                            + c*((y-yo)**2)))
+    return g.ravel()
+
+
+def gaussian2D(tup, amplitude, xo, yo, sigma_x, sigma_y, theta, offset):
+    (x, y) = tup
+    xo = float(xo)
+    yo = float(yo)
+    a = (np.cos(theta)**2)/(2*sigma_x**2) + (np.sin(theta)**2)/(2*sigma_y**2)
+    b = -(np.sin(2*theta))/(4*sigma_x**2) + (np.sin(2*theta))/(4*sigma_y**2)
+    c = (np.sin(theta)**2)/(2*sigma_x**2) + (np.cos(theta)**2)/(2*sigma_y**2)
+    norm = 1/(2*np.pi*np.sqrt(a*c-b*b))
+    g = offset + norm*amplitude*np.exp( - (a*((x-xo)**2) + 2*b*(x-xo)*(y-yo)
+                            + c*((y-yo)**2)))
+    return g.ravel()
+
+
+def rebinND(a, shape):
+    sh = shape[0],a.shape[0]//shape[0],shape[1],a.shape[1]//shape[1]
+    return a.reshape(sh).mean(-1).mean(1)
+
+
+def extract_standard_spec(stdcube, subsample=20, method='boxcar'):
+    """ Extract a spectrum of a standard star from a datacube
+
+    Args:
+        std_cube (`astropy.io.fits.HDUList`_):
+            An HDU list of fits files
+        subsample (int):
+            Number of pixels to subpixellate spectrum when creating mask
+        method (str):
+            Method used to extract standard star spectrum
+
+    Returns:
+        wave (`numpy.ndarray`_): Wavelength of the star.
+        Nlam_star (`numpy.ndarray`_): counts/second/Angstrom
+        Nlam_ivar_star (`numpy.ndarray`_): inverse variance of Nlam_star
+        gpm_star (`numpy.ndarray`_): good pixel mask for Nlam_star
+    """
+    # Extract some information from the HDU list
+    flxcube = stdcube['FLUX'].data.T.copy()
+    varcube = stdcube['VARIANCE'].data.T.copy()
+    numwave = flxcube.shape[2]
+
+    # Setup the WCS
+    stdwcs = wcs.WCS(stdcube['FLUX'].header)
+    wcs_wav = stdwcs.wcs_pix2world(np.vstack((np.zeros(numwave), np.zeros(numwave), np.arange(numwave))).T, 0)
+    wave = wcs_wav[:, 2] * 1.0E10 * units.AA
+
+    # Generate a whitelight image
+    nrmval = np.sum(flxcube != 0.0, axis=2)
+    nrmval[nrmval == 0.0] = 1.0
+    wl_img = np.sum(flxcube, axis=2) / nrmval
+    wlscl = np.max(wl_img)  # Need to make sure the value is of order 1, so it's the same order of magnitude as the other parameters
+
+    # Estimate the centroid and width of the standard star
+    x = np.linspace(0, wl_img.shape[0] - 1, wl_img.shape[0])
+    y = np.linspace(0, wl_img.shape[1] - 1, wl_img.shape[1])
+    xx, yy = np.meshgrid(x, y, indexing='ij')
+    idx_max = np.unravel_index(np.argmax(wl_img), wl_img.shape)
+    initial_guess = (1, idx_max[0], idx_max[1], 2, 2, 0, 0)
+    wlscl = np.max(
+        wl_img)  # Need to make sure the value is of order 1, so it's the same order of magnitude as the other parameters
+    bounds = ([0, 0, 0, 0.5, 0.5, -np.pi, -np.inf],
+              [np.inf, wl_img.shape[0], wl_img.shape[1], wl_img.shape[0], wl_img.shape[1], np.pi, np.inf])
+    popt, pcov = opt.curve_fit(gaussian2D, (xx, yy), wl_img.ravel() / wlscl, bounds=bounds, p0=initial_guess)
+    wid = max(popt[3], popt[4])
+
+    # Setup the coordinates of the mask
+    x = np.linspace(0, flxcube.shape[0] - 1, flxcube.shape[0] * subsample)
+    y = np.linspace(0, flxcube.shape[1] - 1, flxcube.shape[1] * subsample)
+    xx, yy = np.meshgrid(x, y, indexing='ij')
+
+    # Generate a mask
+    newshape = (flxcube.shape[0] * subsample, flxcube.shape[1] * subsample)
+    mask = np.zeros(newshape)
+    nsig = 4  # 4 sigma should be far enough... Note: percentage enclosed for 2D Gaussian = 1-np.exp(-0.5 * nsig**2)
+    ww = np.where((np.sqrt((xx - popt[1]) ** 2 + (yy - popt[2]) ** 2) < nsig * wid))
+    mask[ww] = 1
+    mask = rebinND(mask, (flxcube.shape[0], flxcube.shape[1])).reshape(flxcube.shape[0], flxcube.shape[1], 1)
+
+    # Generate a sky mask
+    newshape = (flxcube.shape[0] * subsample, flxcube.shape[1] * subsample)
+    smask = np.zeros(newshape)
+    nsig = 8  # 8 sigma should be far enough
+    ww = np.where((np.sqrt((xx - popt[1]) ** 2 + (yy - popt[2]) ** 2) < nsig * wid))
+    smask[ww] = 1
+    smask = rebinND(smask, (flxcube.shape[0], flxcube.shape[1])).reshape(flxcube.shape[0], flxcube.shape[1], 1)
+    smask -= mask
+
+    # Subtract the residual sky
+    skymask = (varcube > 0.0) * smask
+    skycube = flxcube * skymask
+    skyspec = skycube.sum(0).sum(0)
+    nrmsky = skymask.sum(0).sum(0)
+    skyspec *= utils.inverse(nrmsky)
+    flxcube -= skyspec.reshape((1, 1, numwave))
+
+    if method == 'boxcar':
+        msgs.info("Extracting a boxcar spectrum of datacube")
+        # Extract boxcar
+        cntmask = (varcube > 0.0) * mask
+        scimask = flxcube * cntmask
+        varmask = varcube * cntmask**2
+        cnt_spec = cntmask.sum(0).sum(0) * utils.inverse(mask.sum())
+        nrmcnt = utils.inverse(cnt_spec)
+        box_flux = scimask.sum(0).sum(0) * nrmcnt
+        box_var = varmask.sum(0).sum(0) * nrmcnt**2
+        box_gpm = np.ones(box_flux.size, dtype=np.bool)
+        # Setup the return values
+        ret_flux, ret_var, ret_gpm = box_flux, box_var, box_gpm
+    elif method == 'gauss2d':
+        # Generate a mask
+        fitmask = (varcube > 0.0) * mask
+        # Estimate the centroid and width of the standard star
+        x = np.linspace(0, flxcube.shape[0] - 1, flxcube.shape[0])
+        y = np.linspace(0, flxcube.shape[1] - 1, flxcube.shape[1])
+        z = np.linspace(0, flxcube.shape[2] - 1, flxcube.shape[2])
+        xx, yy, zz = np.meshgrid(x, y, z, indexing='ij')
+        ww = np.where(fitmask)
+        initial_guess = (1, idx_max[0], idx_max[1], 0.0, 0.0, 2, 2, 0, 0)
+        bounds = ([-np.inf, 0, 0, -np.inf, -np.inf, 0.5, 0.5, -np.pi, -np.inf],
+                  [np.inf,wl_img.shape[0],wl_img.shape[1],np.inf, np.inf, wl_img.shape[0],wl_img.shape[0],np.pi,np.inf])
+        msgs.info("Fitting a 2D Gaussian to the datacube")
+        popt, pcov = opt.curve_fit(gaussian2D_cube, (xx[ww], yy[ww], zz[ww]), flxcube[ww] / wlscl,
+                                   # sigma=np.ones(ww[0].size), bounds=bounds, p0=initial_guess)
+                                   sigma=np.sqrt(varcube[ww]) / wlscl, bounds=bounds, p0=initial_guess)
+        # Subtract off the best-fitting continuum
+        popt[-1] = 0
+        # Generate the best-fitting model to be used as an optimal profile
+        model = gaussian2D_cube((xx, yy, zz), *popt).reshape(flxcube.shape)
+        numim = flxcube.shape[0]*flxcube.shape[1]
+
+        # Optimally extract
+        msgs.info("Optimally extracting...")
+        sciimg = (flxcube*mask).reshape((numim, numwave)).T
+        ivar = utils.inverse((varcube*mask**2).reshape((numim, numwave)).T)
+        optmask = fitmask.reshape((numim, numwave)).T
+        waveimg = np.ones((numwave, numim))  # Just a dummy array - not needed
+        skyimg = np.zeros((numwave, numim))  # Just a dummy array - not needed
+        thismask = np.ones((numwave, numim))  # Just a dummy array - not needed
+        oprof = model.reshape((numim, numwave)).T
+        sobj = specobj.SpecObj('IFU', 'DET01', SLITID=0)
+        extract.extract_optimal(sciimg, ivar, optmask, waveimg, skyimg, thismask, oprof, sobj)
+        opt_flux, opt_var, opt_gpm = sobj.OPT_COUNTS, sobj.OPT_COUNTS_SIG**2, sobj.OPT_MASK
+        # Setup the return values
+        ret_flux, ret_var, ret_gpm = opt_flux, opt_var, opt_gpm
+    elif method == 'optimal':
+        # First do a boxcar along one dimension
+        msgs.info("Collapsing datacube to a 2D image")
+        omask = mask+smask
+        idx_sum = 0
+        cntmask = (varcube > 0.0) * omask
+        scimask = flxcube * cntmask
+        varmask = varcube * cntmask**2
+        cnt_spec = cntmask.sum(idx_sum) * utils.inverse(omask.sum(idx_sum))
+        nrmcnt = utils.inverse(cnt_spec)
+        box_sciimg = scimask.sum(idx_sum) * nrmcnt
+        box_scivar = varmask.sum(idx_sum) * nrmcnt**2
+        box_sciivar = utils.inverse(box_scivar)
+        # Transpose for optimal
+        box_sciimg = box_sciimg.T
+        box_sciivar = box_sciivar.T
+
+        # Prepare for optimal
+        msgs.info("Starting optimal extraction")
+        thismask = np.ones(box_sciimg.shape, dtype=np.bool)
+        nspec, nspat = thismask.shape[0], thismask.shape[1]
+        slit_left = np.zeros(nspec)
+        slit_right = np.ones(nspec)*(nspat-1)
+        tilts = np.outer(np.linspace(0.0,1.0,nspec), np.ones(nspat))
+        waveimg = np.outer(wave.value, np.ones(nspat))
+        global_sky = np.zeros_like(box_sciimg)
+        # Find objects and then extract
+        sobj = findobj_skymask.objs_in_slit(box_sciimg, thismask, slit_left, slit_right)
+        skysub.local_skysub_extract(box_sciimg, box_sciivar, tilts, waveimg, global_sky, thismask, slit_left,
+                             slit_right, sobj, model_noise=False)
+        opt_flux, opt_var, opt_gpm = sobj.OPT_COUNTS[0,:], sobj.OPT_COUNTS_SIG[0,:]**2, sobj.OPT_MASK[0,:]
+        # Setup the return values
+        ret_flux, ret_var, ret_gpm = opt_flux, opt_var, opt_gpm
+
+        ###  OLD VERSION OF OPTIMAL USING 2D PROFILE
+        # # Generate a mask
+        # mask = (mask==1).astype(np.float)
+        # fitmask = (varcube > 0.0) * mask
+        # numim = flxcube.shape[0]*flxcube.shape[1]
+        # wl_img -= np.sum(wl_img[:, :, np.newaxis] * smask)/np.sum(smask)
+        # model = wl_img[:, :, np.newaxis] * mask
+        # # Optimally extract
+        # msgs.info("Optimally extracting...")
+        # sciimg = (flxcube*mask).reshape((numim, numwave)).T
+        # ivar = utils.inverse((varcube*mask**2).reshape((numim, numwave)).T)
+        # optmask = (fitmask > 0).reshape((numim, numwave)).T
+        # waveimg = np.ones((numwave, numim))  # Just a dummy array - not needed
+        # skyimg = np.zeros((numwave, numim))  # Just a dummy array - not needed
+        # thismask = np.ones((numwave, numim))  # Just a dummy array - not needed
+        # oprof = model.repeat(numwave, axis=2).reshape((numim, numwave)).T
+        # sobj = specobj.SpecObj('IFU', 'DET01', SLITID=0)
+        # extract.extract_optimal(sciimg, ivar, optmask, waveimg, skyimg, thismask, oprof, sobj)
+        # opt_flux, opt_var, opt_gpm = sobj.OPT_COUNTS, sobj.OPT_COUNTS_SIG**2, sobj.OPT_MASK
+        # # Setup the return values
+        # ret_flux, ret_var, ret_gpm = opt_flux, opt_var, opt_gpm
+    else:
+        msgs.error("Unknown extraction method: ", method)
+
+    # Convert to counts/s/A
+    arcsecSQ = 3600.0*3600.0*(stdwcs.wcs.cdelt[0]*stdwcs.wcs.cdelt[1])
+    ret_flux *= arcsecSQ
+    ret_var *= arcsecSQ**2
+    # Return the box extraction results
+    return wave, ret_flux, utils.inverse(ret_var), ret_gpm
 
 
 def make_whitelight_fromref(all_ra, all_dec, all_wave, all_sci, all_wghts, all_idx, dspat, ref_filename):
@@ -540,9 +764,98 @@ def compute_weights(all_ra, all_dec, all_wave, all_sci, all_ivar, all_idx, white
         ww = (all_idx == ff)
         all_wghts[ww] = interp1d(wave_spec, weights[:, ff], kind='cubic',
                                  bounds_error=False, fill_value="extrapolate")(all_wave[ww])
-
     msgs.info("Optimal weighting complete")
     return all_wghts
+
+
+def generate_cube_ngp(outfile, hdr, all_sci, all_ivar, all_wghts, pix_coord, bins,
+                      overwrite=False, blaze_wave=None, blaze_spec=None, fluxcal=False,
+                      specname="PYP_SPEC", debug=False):
+    """
+    Save a datacube using the Nearest Grid Point (NGP) algorithm.
+
+    Args:
+        outfile (`str`):
+            Filename to be used to save the datacube
+        hdr (`astropy.io.fits.header_`):
+            Header of the output datacube (must contain WCS)
+        all_sci (`numpy.ndarray`_):
+            1D flattened array containing the counts of each pixel from all spec2d files
+        all_ivar (`numpy.ndarray`_):
+            1D flattened array containing the inverse variance of each pixel from all spec2d files
+        all_wghts (`numpy.ndarray`_):
+            1D flattened array containing the weights of each pixel to be used in the combination
+        pix_coord (`numpy.ndarray`_):
+            The NGP pixel coordinates corresponding to the RA,DEC,WAVELENGTH of each individual
+            pixel in the processed spec2d frames. After setting up an astropy WCS, pix_coord is
+            returned by the function: `astropy.wcs.WCS.wcs_world2pix_`
+        bins (tuple):
+            A 3-tuple (x,y,z) containing the histogram bin edges in x,y spatial and z wavelength coordinates    :param overwrite:
+        blaze_wave (`numpy.ndarray`_):
+            Wavelength array of the spectral blaze function
+        blaze_spec (`numpy.ndarray`_):
+            Spectral blaze function
+        fluxcal (bool):
+            Are the data flux calibrated?
+        specname (str):
+            Name of the spectrograph
+        debug (bool):
+            Debug the code by writing out a residuals cube?
+    """
+    # Add the unit of flux to the header
+    if fluxcal:
+        hdr['FLUXUNIT'] = (PYPEIT_FLUX_SCALE, "Flux units -- erg/s/cm^2/Angstrom/arcsec^2")
+    else:
+        hdr['FLUXUNIT'] = (1, "Flux units -- counts/s/Angstrom/arcsec^2")
+    # from matplotlib import pyplot as plt
+    # embed()
+    # assert(False)
+    # mskvals = np.load("mskvals.npy")
+    # resid = all_sci*np.sqrt(all_ivar)*mskvals
+    # resid = resid[resid!=0]
+    # nrm, _, _ = plt.hist(resid.flatten(), bins=100)
+    # rx = np.linspace(-10,10,1000)
+    # ry = np.max(nrm)*np.exp(-0.5*rx**2)
+    # plt.plot(rx, ry, 'k-')
+    # plt.show()
+    #
+    # ww = np.where(mskvals)
+    # new_pix_coord = [pix_coord[0][ww], pix_coord[1][ww], pix_coord[2][ww]]
+    # # Use NGP to generate the cube - this ensures errors between neighbouring voxels are not correlated
+    # datacube, edges = np.histogramdd(new_pix_coord, bins=bins, weights=all_sci[ww] * all_wghts[ww])
+    # norm, edges = np.histogramdd(new_pix_coord, bins=bins, weights=all_wghts[ww])
+    # norm_cube = utils.inverse(norm)
+    # datacube *= norm_cube
+    # # Create the variance cube, including weights
+    # msgs.info("Generating variance cube")
+    # all_var = utils.inverse(all_ivar)
+    # var_cube, edges = np.histogramdd(new_pix_coord, bins=bins, weights=all_var[ww] * all_wghts[ww]**2)
+    # var_cube *= norm_cube**2
+
+    # Use NGP to generate the cube - this ensures errors between neighbouring voxels are not correlated
+    datacube, edges = np.histogramdd(pix_coord, bins=bins, weights=all_sci * all_wghts)
+    norm, edges = np.histogramdd(pix_coord, bins=bins, weights=all_wghts)
+    norm_cube = utils.inverse(norm)
+    datacube *= norm_cube
+    # Create the variance cube, including weights
+    msgs.info("Generating variance cube")
+    all_var = utils.inverse(all_ivar)
+    var_cube, edges = np.histogramdd(pix_coord, bins=bins, weights=all_var * all_wghts**2)
+    var_cube *= norm_cube**2
+
+    # Save the datacube
+    if debug:
+        datacube_resid, edges = np.histogramdd(pix_coord, bins=bins, weights=all_sci*np.sqrt(all_ivar))
+        norm, edges = np.histogramdd(pix_coord, bins=bins)
+        norm_cube = utils.inverse(norm)
+        outfile_resid = "datacube_resid.fits"
+        msgs.info("Saving datacube as: {0:s}".format(outfile_resid))
+        hdu = fits.PrimaryHDU((datacube_resid*norm_cube).T, header=hdr)
+        hdu.writeto(outfile_resid, overwrite=overwrite)
+
+    msgs.info("Saving datacube as: {0:s}".format(outfile))
+    final_cube = DataCube(datacube.T, var_cube.T, specname, blaze_wave, blaze_spec, fluxed=fluxcal)
+    final_cube.to_file(outfile, hdr=hdr, overwrite=overwrite)
 
 
 def coadd_cube(files, spectrograph=None, parset=None, overwrite=False):
@@ -583,61 +896,146 @@ def coadd_cube(files, spectrograph=None, parset=None, overwrite=False):
     if parset is None:
         parset = spec.default_pypeit_par()
     cubepar = parset['reduce']['cube']
-
-    # Check the output file
-    outfile = cubepar['output_filename'] if ".fits" in cubepar['output_filename'] else cubepar['output_filename']+".fits"
-    out_whitelight = outfile.replace(".fits", "_whitelight.fits")
-    if os.path.exists(outfile) and not overwrite:
-        msgs.error("Output filename already exists:"+msgs.newline()+outfile)
-    elif os.path.exists(out_whitelight) and cubepar['save_whitelight'] and not overwrite:
-        msgs.error("Output filename already exists:"+msgs.newline()+out_whitelight)
-    # Check the reference cube and image exist, if requested
-    if cubepar['standard_cube'] is not None:
-        if not os.path.exists(cubepar['standard_cube']):
-            msgs.error("Standard cube does not exist:" + msgs.newline() + cubepar['reference_cube'])
-        stdcube = fits.open(cubepar['standard_cube'])
-        msgs.info("Reference scale loaded from file: {0:s}".format(cubepar['standard_cube']))
-    if cubepar['reference_image'] is not None:
-        if not os.path.exists(cubepar['reference_image']):
-            msgs.error("Reference cube does not exist:" + msgs.newline() + cubepar['reference_image'])
-    if cubepar['flux_calibrate']:
-        msgs.error("Flux calibration is not currently implemented" + msgs.newline() +
-                   "Please set 'flux_calibrate = False'")
+    flatpar = parset['calibrations']['flatfield']
 
     # prep
     numfiles = len(files)
     combine = cubepar['combine']
 
+    # Check the output files don't exist
+    outfile = cubepar['output_filename'] if ".fits" in cubepar['output_filename'] else cubepar['output_filename'] + ".fits"
+    out_whitelight = outfile.replace(".fits", "_whitelight.fits")
+    if combine:
+        if os.path.exists(outfile) and not overwrite:
+            msgs.error("Output filename already exists:"+msgs.newline()+outfile)
+        if os.path.exists(out_whitelight) and cubepar['save_whitelight'] and not overwrite:
+            msgs.error("Output filename already exists:"+msgs.newline()+out_whitelight)
+    else:
+        for ff in range(numfiles):
+            outfile = files[ff].replace("spec2d_", "spec3d_")
+            out_whitelight = outfile.replace(".fits", "_whitelight.fits")
+            if os.path.exists(outfile) and not overwrite:
+                msgs.error("Output filename already exists:" + msgs.newline() + outfile)
+            if os.path.exists(out_whitelight) and cubepar['save_whitelight'] and not overwrite:
+                msgs.error("Output filename already exists:" + msgs.newline() + out_whitelight)
+
+    # Check the reference cube and image exist, if requested
+    fluxcal = False
+    blaze_wave, blaze_spec = None, None
+    blaze_spline, flux_spline = None, None
+    if cubepar['standard_cube'] is not None:
+        if not os.path.exists(cubepar['standard_cube']):
+            msgs.error("Standard cube does not exist:" + msgs.newline() + cubepar['reference_cube'])
+        fluxcal = True
+        senspar = parset['sensfunc']
+        # Load the standard star cube and retrieve its RA + DEC
+        stdcube = fits.open(cubepar['standard_cube'])
+        star_ra, star_dec = stdcube[1].header['CRVAL1'], stdcube[1].header['CRVAL2']
+        # Extract the information about the blaze
+        blaze_wave, blaze_spec = stdcube['BLAZE_WAVE'].data, stdcube['BLAZE_SPEC'].data
+        blaze_spline = interp1d(blaze_wave, blaze_spec, kind='linear', bounds_error=False, fill_value="extrapolate")
+        # Extract a spectrum of the standard star
+        wave, Nlam_star, Nlam_ivar_star, gpm_star = extract_standard_spec(stdcube)
+        # Read in some information above the standard star
+        std_dict = get_standard_spectrum(star_type=senspar['star_type'],
+                                         star_mag=senspar['star_mag'],
+                                         ra=star_ra, dec=star_dec)
+        # Calculate the sensitivity curve
+        zeropoint_data, zeropoint_data_gpm, zeropoint_fit, zeropoint_fit_gpm =\
+            fit_zeropoint(wave.value, Nlam_star, Nlam_ivar_star, gpm_star, std_dict,
+                          mask_abs_lines=senspar['mask_abs_lines'], balm_mask_wid=senspar['UVIS']['balm_mask_wid'],
+                          nresln=senspar['UVIS']['nresln'], resolution=senspar['UVIS']['resolution'],
+                          trans_thresh=senspar['UVIS']['trans_thresh'], polyorder=senspar['polyorder'],
+                          polycorrect=senspar['UVIS']['polycorrect'], polyfunc=senspar['UVIS']['polyfunc'])
+        wgd = np.where(zeropoint_fit_gpm)
+        sens = np.power(10.0, -0.4 * (zeropoint_fit[wgd] - ZP_UNIT_CONST)) / np.square(wave[wgd])
+        flux_spline = interp1d(wave[wgd], sens, kind='linear', bounds_error=False, fill_value="extrapolate")
+    if cubepar['reference_image'] is not None:
+        if not os.path.exists(cubepar['reference_image']):
+            msgs.error("Reference cube does not exist:" + msgs.newline() + cubepar['reference_image'])
+
+    # Initialise arrays for storage
     all_ra, all_dec, all_wave = np.array([]), np.array([]), np.array([])
     all_sci, all_ivar, all_idx, all_wghts = np.array([]), np.array([]), np.array([]), np.array([])
     all_wcs = []
-    dspat = None if cubepar['spatial_delta'] is None else  cubepar['spatial_delta']/3600.0  # binning size on the sky (/3600 to convert to degrees)
+    dspat = None if cubepar['spatial_delta'] is None else cubepar['spatial_delta']/3600.0  # binning size on the sky (/3600 to convert to degrees)
     dwv = cubepar['wave_delta']       # binning size in wavelength direction (in Angstroms)
     wave_ref = None
     whitelight_img = None  # This is the whitelight image based on all input spec2d frames
     weights = np.ones(numfiles)  # Weights to use when combining cubes
     locations = parset['calibrations']['alignment']['locations']
+    flat_splines = dict()   # A dictionary containing the splines of the flatfield
+    # Load the scaleimg frame for the scale correction
+    if cubepar['scale_corr'] is not None:
+        msgs.info("Loading scale image for relative spectral illumination correction:" +
+                  msgs.newline() + cubepar['scale_corr'])
+        try:
+            spec2DObj = spec2dobj.Spec2DObj.from_file(cubepar['scale_corr'], detname)
+            relScaleImg = spec2DObj.scaleimg
+        except:
+            msgs.warn("Could not load scaleimg from spec2d file:" + msgs.newline() + cubepar['scale_corr'] +
+                      "scale correction will not be performed")
+            cubepar['scale_corr'] = None
+    # Load all spec2d files and prepare the data for making a datacube
     for ff, fil in enumerate(files):
         # Load it up
         spec2DObj = spec2dobj.Spec2DObj.from_file(fil, detname)
         detector = spec2DObj.detector
         flexure = None  #spec2DObj.sci_spat_flexure
 
+        # Load the header
+        hdr = fits.open(fil)[0].header
+
+        # Get the exposure time
+        exptime = hdr['EXPTIME']
+
         # Setup for PypeIt imports
         msgs.reset(verbosity=2)
 
         # Extract the information
-        sciimg = (spec2DObj.sciimg-spec2DObj.skymodel)  # Subtract sky
-        ivar = spec2DObj.ivarraw
+        relscl = spec2DObj.scaleimg/relScaleImg
+        sciimg = (spec2DObj.sciimg-spec2DObj.skymodel)*relscl  # Subtract sky
+        ivar = spec2DObj.ivarraw / relscl**2
         waveimg = spec2DObj.waveimg
         bpmmask = spec2DObj.bpmmask
 
         # Grab the slit edges
         slits = spec2DObj.slits
 
+        # embed()
+        # assert(False)
+        # from pypeit.core import skysub
+        # skyregtxt = ":30,60:"  # BB08
+        # slits_left, slits_right, _ = slits.select_edges(initial=True, flexure=flexure)
+        # maxslitlength = np.max(slits_right - slits_left)
+        # # Get the regions
+        # status, regions = skysub.read_userregions(skyregtxt, slits.nslits, maxslitlength)
+        # # Generate image
+        # skymask_init = skysub.generate_mask("IFU", regions, slits, slits_left, slits_right, spat_flexure=flexure)
+        # resid = sciimg*np.sqrt(ivar)*skymask_init*(bpmmask==0)*(ivar!=0)
+        # resid = resid[resid!=0.0]
+        # nrm, _, _ = plt.hist(resid.flatten(), bins=100)
+        # rx = np.linspace(-10,10,1000)
+        # ry = np.max(nrm)*np.exp(-0.5*rx**2)
+        # plt.plot(rx, ry, 'k-')
+        # plt.show()
+
         wave0 = waveimg[waveimg != 0.0].min()
-        diff = waveimg[1:, :] - waveimg[:-1, :]
-        dwv = float(np.median(diff[diff != 0.0]))
+        # Calculate the delta wave in every pixel on the slit
+        waveimp = np.roll(waveimg, 1, axis=0)
+        waveimn = np.roll(waveimg, -1, axis=0)
+        dwaveimg = np.zeros_like(waveimg)
+        # All good pixels
+        wnz = np.where((waveimg!=0)&(waveimp!=0))
+        dwaveimg[wnz] = np.abs(waveimg[wnz]-waveimp[wnz])
+        # All bad pixels
+        wnz = np.where((waveimg!=0)&(waveimp==0))
+        dwaveimg[wnz] = np.abs(waveimg[wnz]-waveimn[wnz])
+        # All endpoint pixels
+        dwaveimg[0, :] = np.abs(waveimg[0, :] - waveimn[0, :])
+        dwaveimg[-1, :] = np.abs(waveimg[-1, :] - waveimp[-1, :])
+        dwv = np.median(dwaveimg[dwaveimg != 0.0]) if cubepar['wave_delta'] is None else cubepar['wave_delta']
+
         msgs.info("Using wavelength solution: wave0={0:.3f}, dispersion={1:.3f} Angstrom/pixel".format(wave0, dwv))
 
         msgs.info("Constructing slit image")
@@ -654,13 +1052,14 @@ def coadd_cube(files, spectrograph=None, parset=None, overwrite=False):
         slscl = spec.get_meta_value([spec2DObj.head0], 'slitwid')
         if dspat is None:
             dspat = max(pxscl, slscl)
-        elif max(pxscl, slscl) > dspat:
-            dspat = max(pxscl, slscl)
+        if pxscl > dspat:
+            msgs.warn("Spatial scale requested ({0:f}'') is less than the pixel scale ({1:f}'')".format(dspat,pxscl))
+        if pxscl > dspat:
+            msgs.warn("Spatial scale requested ({0:f}'') is less than the slicer scale ({1:f}'')".format(dspat, slscl))
 
         # Loading the alignments frame for these data
         astrometric = cubepar['astrometric']
         msgs.info("Loading alignments")
-        hdr = fits.open(fil)[0].header
         alignfile = masterframe.construct_file_name(alignframe.Alignments, hdr['TRACMKEY'], master_dir=hdr['PYPMFDIR'])
         alignments = None
         if os.path.exists(alignfile) and cubepar['astrometric']:
@@ -703,6 +1102,37 @@ def coadd_cube(files, spectrograph=None, parset=None, overwrite=False):
         wave_ext = waveimg[onslit_gpm].copy()
         flux_ext = sciimg[onslit_gpm].copy()
         ivar_ext = ivar[onslit_gpm].copy()
+        dwav_ext = dwaveimg[onslit_gpm].copy()
+
+        # Correct for sensitivity as a function of grating angle
+        # (this assumes the spectrum of the flatfield lamp has the same shape for all setups)
+        flatfile = masterframe.construct_file_name(flatfield.FlatImages, hdr['FLATMKEY'], master_dir=hdr['PYPMFDIR'])
+        if cubepar['grating_corr'] and flatfile not in flat_splines.keys():
+            msgs.info("Calculating relative sensitivity for grating correction")
+            flatimages = flatfield.FlatImages.from_file(flatfile)
+            flatframe = flatimages.pixelflat_model
+            flatframe /= flatimages.fit2illumflat(slits, frametype='pixel', initial=True, flexure_shift=flexure)
+            # Calculate the relative scale
+            scale_model = flatfield.illum_profile_spectral(flatframe, waveimg, slits,
+                                                           slit_illum_ref_idx=flatpar['slit_illum_ref_idx'], model=None,
+                                                           skymask=None, trim=flatpar['slit_trim'], flexure=flexure)
+            # Apply the relative scale and generate a 1D "spectrum"
+            onslit = waveimg != 0
+            wavebins = np.linspace(np.min(waveimg[onslit]), np.max(waveimg[onslit]), slits.nspec)
+            hist, edge = np.histogram(waveimg[onslit], bins=wavebins, weights=flatframe[onslit]/scale_model[onslit])
+            cntr, edge = np.histogram(waveimg[onslit], bins=wavebins)
+            cntr = cntr.astype(np.float)
+            norm = (cntr != 0) / (cntr + (cntr == 0))
+            spec_spl = hist * norm
+            wave_spl = 0.5 * (wavebins[1:] + wavebins[:-1])
+            flat_splines[flatfile] = interp1d(wave_spl, spec_spl, kind='linear',
+                                              bounds_error=False, fill_value="extrapolate")
+            # Check if a reference blaze spline exists (either from a standard star if fluxing or from a previous
+            # exposure in this for loop)
+            if blaze_spline is None:
+                blaze_wave, blaze_spec = wave_spl, spec_spl
+                blaze_spline = interp1d(wave_spl, spec_spl, kind='linear',
+                                        bounds_error=False, fill_value="extrapolate")
 
         # Perform extinction correction
         msgs.info("Applying extinction correction")
@@ -713,79 +1143,130 @@ def coadd_cube(files, spectrograph=None, parset=None, overwrite=False):
         # extinction_correction requires the wavelength is sorted
         wvsrt = np.argsort(wave_ext)
         ext_corr = extinction_correction(wave_ext[wvsrt] * units.AA, airmass, extinct)
+        # Grating correction
+        grat_corr = 1.0
+        if cubepar['grating_corr']:
+            msgs.info("Calculating the grating correction")
+            grat_corr_tmp = flat_splines[flatfile](wave_ext[wvsrt]) / blaze_spline(wave_ext[wvsrt])
+            # Fit a low order polynomial to this correction
+            minw, maxw = max(np.min(wave_spl), np.min(blaze_wave)), max(np.min(wave_spl), np.max(blaze_wave))
+            wblz = np.where((wave_ext[wvsrt] > minw) & (wave_ext[wvsrt] < maxw))
+            wave_corr = (wave_ext[wvsrt] - minw) / (maxw - minw)
+            coeff_gratcorr = np.polyfit(wave_corr[wblz], grat_corr_tmp[wblz], 2)
+            grat_corr = np.polyval(coeff_gratcorr, wave_corr)
+        # Sensitivity function
+        sens_func = 1.0
+        if fluxcal:
+            msgs.info("Calculating the sensitivity function")
+            sens_func = flux_spline(wave_ext[wvsrt])
+        # Convert the flux_sav to counts/s,  correct for the relative sensitivity of different setups
+        ext_corr *= sens_func / (exptime * grat_corr)
         # Correct for extinction
         flux_sav = flux_ext[wvsrt] * ext_corr
         ivar_sav = ivar_ext[wvsrt] / ext_corr ** 2
+
+        # Convert units to Counts/s/Ang/arcsec2
+        scl_units = dwav_ext[wvsrt] * 3600.0 * 3600.0 * (frame_wcs.wcs.cdelt[0] * frame_wcs.wcs.cdelt[1])
+        flux_sav /= scl_units
+        ivar_sav *= scl_units ** 2
+
         # sort back to the original ordering
         resrt = np.argsort(wvsrt)
+        numpix = raimg[onslit_gpm].size
 
         # Calculate the weights relative to the zeroth cube
-        if ff != 0:
-            weights[ff] = np.median(flux_sav[resrt]*np.sqrt(ivar_sav[resrt]))**2
+        # TODO :: This is dodgy when there is mostly sky...
+        weights[ff] = np.median(flux_sav[resrt]*np.sqrt(ivar_sav[resrt]))**2
+
+        # If individual frames are to be output, there's no need to store information, just make the cubes now
+        if not combine:
+            outfile = fil.replace("spec2d_", "spec3d_")
+            if cubepar['save_whitelight']:
+                # Generate individual whitelight images of each spec2d file
+                out_whitelight = outfile.replace(".fits", "_whitelight.fits")
+                whitelight_img, _, wlwcs = make_whitelight(raimg[onslit_gpm], decimg[onslit_gpm], wave_ext,
+                                                           flux_sav[resrt], np.ones(numpix), np.zeros(numpix), dspat)
+                msgs.info("Saving white light image as: {0:s}".format(out_whitelight))
+                img_hdu = fits.PrimaryHDU(whitelight_img.T, header=wlwcs.to_header())
+                img_hdu.writeto(out_whitelight, overwrite=overwrite)
+
+            slitlength = int(np.round(np.median(slits.get_slitlengths(initial=True, median=True))))
+            numwav = int((np.max(waveimg) - wave0) / dwv)
+            bins = spec.get_datacube_bins(slitlength, minmax, numwav)
+            msgs.info("Generating pixel coordinates")
+            pix_coord = frame_wcs.wcs_world2pix(np.vstack((raimg[onslit_gpm], decimg[onslit_gpm], wave_ext * 1.0E-10)).T, 0)
+            hdr = frame_wcs.to_header()
+            generate_cube_ngp(outfile, hdr, flux_sav[resrt], ivar_sav[resrt], np.ones(numpix), pix_coord, bins,
+                              overwrite=overwrite, blaze_wave=blaze_wave, blaze_spec=blaze_spec,
+                              fluxcal=fluxcal, specname=specname)
+            continue
 
         # Store the information
-        numpix = raimg[onslit_gpm].size
         all_ra = np.append(all_ra, raimg[onslit_gpm].copy())
         all_dec = np.append(all_dec, decimg[onslit_gpm].copy())
         all_wave = np.append(all_wave, wave_ext.copy())
         all_sci = np.append(all_sci, flux_sav[resrt].copy())
         all_ivar = np.append(all_ivar, ivar_sav[resrt].copy())
         all_idx = np.append(all_idx, ff*np.ones(numpix))
-        all_wghts = np.append(all_wghts, weights[ff]*np.ones(numpix))
+        all_wghts = np.append(all_wghts, weights[ff]*np.ones(numpix)/weights[0])
+
+    # No need to continue if we are not combining frames
+    if not combine:
+        return
 
     # Grab cos(dec) for convenience
     cosdec = np.cos(np.mean(all_dec) * np.pi / 180.0)
 
-    # Register spatial offsets between all frames if several frames are being combined
-    if combine:
-        # Check if a reference whitelight image should be used to register the offsets
-        if cubepar["reference_image"] is None:
-            # Generate white light images
-            whitelight_imgs, _, _ = make_whitelight(all_ra, all_dec, all_wave, all_sci, all_wghts, all_idx, dspat)
-            # ref_idx will be the index of the cube with the highest S/N
-            ref_idx = np.argmax(weights)
-            reference_image = whitelight_imgs[:, :, ref_idx].copy()
-            msgs.info("Calculating spatial translation of each cube relative to cube #{0:d})".format(ref_idx+1))
-        else:
-            ref_idx = -1  # Don't use an index
-            # Load reference information
-            reference_image, whitelight_imgs, wlwcs = \
-                make_whitelight_fromref(all_ra, all_dec, all_wave, all_sci, all_wghts, all_idx, dspat,
-                                        cubepar['reference_image'])
-            msgs.info("Calculating the spatial translation of each cube relative to user-defined 'reference_image'")
-        # Calculate the image offsets - check the reference is a zero shift
-        ra_shift_ref, dec_shift_ref = calculate_image_offset(reference_image.copy(), reference_image.copy())
-        for ff in range(numfiles):
-            # Don't correlate the reference image with itself
-            if ff == ref_idx:
-                continue
-            # Calculate the shift
-            ra_shift, dec_shift = calculate_image_offset(whitelight_imgs[:, :, ff], reference_image.copy())
-            # Convert to reference
-            ra_shift -= ra_shift_ref
-            dec_shift -= dec_shift_ref
-            # Convert pixel shift to degress shift
-            ra_shift *= dspat/cosdec
-            dec_shift *= dspat
-            msgs.info("Spatial shift of cube #{0:d}: RA, DEC (arcsec) = {1:+0.3f}, {2:+0.3f}".format(ff+1, ra_shift*3600.0, dec_shift*3600.0))
-            # Apply the shift
-            all_ra[all_idx == ff] += ra_shift
-            all_dec[all_idx == ff] += dec_shift
+    # Register spatial offsets between all frames
+    # Check if a reference whitelight image should be used to register the offsets
+    if cubepar["reference_image"] is None:
+        # Generate white light images
+        whitelight_imgs, _, _ = make_whitelight(all_ra, all_dec, all_wave, all_sci, all_wghts, all_idx, dspat)
+        # ref_idx will be the index of the cube with the highest S/N
+        ref_idx = np.argmax(weights)
+        reference_image = whitelight_imgs[:, :, ref_idx].copy()
+        msgs.info("Calculating spatial translation of each cube relative to cube #{0:d})".format(ref_idx+1))
+    else:
+        ref_idx = -1  # Don't use an index
+        # Load reference information
+        reference_image, whitelight_imgs, wlwcs = \
+            make_whitelight_fromref(all_ra, all_dec, all_wave, all_sci, all_wghts, all_idx, dspat,
+                                    cubepar['reference_image'])
+        msgs.info("Calculating the spatial translation of each cube relative to user-defined 'reference_image'")
 
-        # Generate a white light image of *all* data
-        msgs.info("Generating global white light image")
-        if cubepar["reference_image"] is None:
-            whitelight_img, _, wlwcs = make_whitelight(all_ra, all_dec, all_wave, all_sci, all_wghts,
-                                                       np.zeros(all_ra.size), dspat)
-        else:
-            _, whitelight_img, wlwcs = \
-                make_whitelight_fromref(all_ra, all_dec, all_wave, all_sci, all_wghts, np.zeros(all_ra.size),
-                                        dspat, cubepar['reference_image'])
+    # Calculate the image offsets - check the reference is a zero shift
+    ra_shift_ref, dec_shift_ref = calculate_image_offset(reference_image.copy(), reference_image.copy())
+    for ff in range(numfiles):
+        # Don't correlate the reference image with itself
+        if ff == ref_idx:
+            continue
+        # Calculate the shift
+        ra_shift, dec_shift = calculate_image_offset(whitelight_imgs[:, :, ff], reference_image.copy())
+        # Convert to reference
+        ra_shift -= ra_shift_ref
+        dec_shift -= dec_shift_ref
+        # Convert pixel shift to degress shift
+        ra_shift *= dspat/cosdec
+        dec_shift *= dspat
+        msgs.info("Spatial shift of cube #{0:d}: RA, DEC (arcsec) = {1:+0.3f}, {2:+0.3f}".format(ff+1, ra_shift*3600.0, dec_shift*3600.0))
+        # Apply the shift
+        all_ra[all_idx == ff] += ra_shift
+        all_dec[all_idx == ff] += dec_shift
 
-        # Calculate the relative spectral weights of all pixels
-        all_wghts = compute_weights(all_ra, all_dec, all_wave, all_sci, all_ivar, all_idx,
-                                    whitelight_img[:, :, 0], dspat, dwv,
-                                    relative_weights=cubepar['relative_weights'])
+    # Generate a white light image of *all* data
+    msgs.info("Generating global white light image")
+    if cubepar["reference_image"] is None:
+        whitelight_img, _, wlwcs = make_whitelight(all_ra, all_dec, all_wave, all_sci, all_wghts,
+                                                   np.zeros(all_ra.size), dspat)
+    else:
+        _, whitelight_img, wlwcs = \
+            make_whitelight_fromref(all_ra, all_dec, all_wave, all_sci, all_wghts, np.zeros(all_ra.size),
+                                    dspat, cubepar['reference_image'])
+
+    # Calculate the relative spectral weights of all pixels
+    all_wghts = compute_weights(all_ra, all_dec, all_wave, all_sci, all_ivar, all_idx,
+                                whitelight_img[:, :, 0], dspat, dwv,
+                                relative_weights=cubepar['relative_weights'])
 
     # Check if a whitelight image should be saved
     if cubepar['save_whitelight']:
@@ -813,6 +1294,7 @@ def coadd_cube(files, spectrograph=None, parset=None, overwrite=False):
     wav_min = cubepar['wave_min'] if cubepar['wave_min'] is not None else np.min(all_wave)
     wav_max = cubepar['wave_max'] if cubepar['wave_max'] is not None else np.max(all_wave)
     if cubepar['wave_delta'] is not None: dwv = cubepar['wave_delta']
+
     # Generate a master WCS to register all frames
     coord_min = [ra_min, dec_min, wav_min]
     coord_dlt = [dspat, dspat, dwv]
@@ -827,52 +1309,20 @@ def coadd_cube(files, spectrograph=None, parset=None, overwrite=False):
               msgs.newline() + "-" * 40)
 
     # Generate the output binning
-    if combine:
-        numra = int((ra_max-ra_min) * cosdec / dspat)
-        numdec = int((dec_max-dec_min)/dspat)
-        numwav = int((wav_max-wav_min)/dwv)
-        xbins = np.arange(1+numra)-0.5
-        ybins = np.arange(1+numdec)-0.5
-        spec_bins = np.arange(1+numwav)-0.5
-    else:
-        slitlength = int(np.round(np.median(slits.get_slitlengths(initial=True, median=True))))
-        numwav = int((np.max(waveimg) - wave0) / dwv)
-        xbins, ybins, spec_bins = spec.get_datacube_bins(slitlength, minmax, numwav)
+    numra = int((ra_max-ra_min) * cosdec / dspat)
+    numdec = int((dec_max-dec_min)/dspat)
+    numwav = int((wav_max-wav_min)/dwv)
+    xbins = np.arange(1+numra)-0.5
+    ybins = np.arange(1+numdec)-0.5
+    spec_bins = np.arange(1+numwav)-0.5
+    bins = (xbins, ybins, spec_bins)
 
     # Make the cube
     msgs.info("Generating pixel coordinates")
-    if combine:
-        pix_coord = masterwcs.wcs_world2pix(all_ra, all_dec, all_wave * 1.0E-10, 0)
-        hdr = masterwcs.to_header()
-    else:
-        pix_coord = wcs.wcs_world2pix(np.vstack((all_ra, all_dec, all_wave*1.0E-10)).T, 0)
-        hdr = frame_wcs.to_header()
+    pix_coord = masterwcs.wcs_world2pix(all_ra, all_dec, all_wave * 1.0E-10, 0)
+    hdr = masterwcs.to_header()
 
     # Find the NGP coordinates for all input pixels
     msgs.info("Generating data cube")
-    bins = (xbins, ybins, spec_bins)
-    datacube, edges = np.histogramdd(pix_coord, bins=bins, weights=all_sci*all_wghts)
-    norm, edges = np.histogramdd(pix_coord, bins=bins, weights=all_wghts)
-    norm_cube = (norm > 0) / (norm + (norm == 0))
-    datacube *= norm_cube
-    # Create the variance cube, including weights
-    msgs.info("Generating variance cube")
-    all_var = (all_ivar > 0) / (all_ivar + (all_ivar == 0))
-    var_cube, edges = np.histogramdd(pix_coord, bins=bins, weights=all_var * all_wghts**2)
-    var_cube *= norm_cube**2
-
-    # Save the datacube
-    debug = False
-    if debug:
-        datacube_resid, edges = np.histogramdd(pix_coord, bins=(xbins, ybins, spec_bins), weights=all_sci*np.sqrt(all_ivar))
-        norm, edges = np.histogramdd(pix_coord, bins=(xbins, ybins, spec_bins))
-        norm_cube = (norm > 0) / (norm + (norm == 0))
-        outfile = "datacube_resid.fits"
-        msgs.info("Saving datacube as: {0:s}".format(outfile))
-        hdu = fits.PrimaryHDU((datacube_resid*norm_cube).T, header=masterwcs.to_header())
-        hdu.writeto(outfile, overwrite=overwrite)
-
-    msgs.info("Saving datacube as: {0:s}".format(outfile))
-    final_cube = DataCube(datacube.T, var_cube.T, specname, fluxed=cubepar['flux_calibrate'])
-    final_cube.to_file(outfile, hdr=hdr, overwrite=overwrite)
-
+    generate_cube_ngp(outfile, hdr, all_sci, all_ivar, all_wghts, pix_coord, bins, overwrite=overwrite,
+                      blaze_wave=blaze_wave, blaze_spec=blaze_spec, fluxcal=fluxcal, specname=specname)